+++ eslint.config.mjs
@@ -6,4 +6,10 @@ import tseslint from 'typescript-eslint';
 export default tseslint.config(
   eslint.configs.recommended,
   ...tseslint.configs.recommended,
+  {
+    ignores: [
+      "*.config.js",
+      "build/**/*" // ignores all content under the 'build/' folder
+    ]
+  }
 );
