--- eslintrc.json
@@ -1,30 +1,0 @@
-{
-    "env": {
-        "browser": false,
-        "es6": true,
-        "node": true
-    },
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "project": "tsconfig.json",
-        "sourceMap": "module",
-        "ecmaVersion": "2020"
-    },
-    "plugins": [
-        "@typescript-eslint",
-        "jest"
-    ],
-    "extends": [
-        "eslint:recommended",
-        "plugin:@typescript-eslint/recommended",
-        "plugin:jest/recommended",
-        "prettier"
-    ],
-    "rules": {
-        // The following rules is enabled only to supplement the inline suppression
-        // exammples, and because it is not a recommended rule, you should either
-        // disable it, or understand what it enforces
-        // https://typescript-eslint.io/rules/explicit-return-type/
-        "@typescript-eslint/explicit-function-return-type": "warn"
-    }
-}
