--- jest.config.js
@@ -1,20 +1,0 @@
-const config = {
-    testEnvironment: 'node',
-    preset: 'ts-jest/presets/default-esm',
-    transform: {
-      '^.+\\.m?[tj]s?$': ['ts-jest', { useESM: false }],
-    },
-    moduleNameMapper: {
-      '^(\\.{1,2}/.*)\\.(m)?js$': '$1',
-    },
-    testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.(m)?ts$',
-    coverageDirectory: 'coverage',
-    collectCoverageFrom: [
-      'src/**/*.ts',
-      'src/**/*.mts',
-      '!src/**/*.d.ts',
-      '!src/**/*.d.mts',
-    ],
-  };
-  
-  module.exports = config;
