+++ jest.config.ts
@@ -1,0 +1,23 @@
+import { Config } from "jest";
+
+const config: Config = {
+  testEnvironment: 'node',
+  preset: 'ts-jest/presets/default-esm',
+  transform: {
+    '^.+\\.m?[tj]s?$': ['ts-jest', { useESM: false }],
+  },
+  moduleNameMapper: {
+    '^(\\.{1,2}/.*)\\.(m)?js$': '$1',
+  },
+  testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.(m)?ts$',
+  coverageDirectory: 'coverage',
+  moduleDirectories: ["node_modules", "src"],
+  collectCoverageFrom: [
+    'src/**/*.ts',
+    'src/**/*.mts',
+    '!src/**/*.d.ts',
+    '!src/**/*.d.mts',
+  ],
+};
+
+export default config;
