+++ package.json
@@ -10,15 +10,19 @@
   "license": "MIT",
   "private": false,
   "scripts": {
+    "pretest": "npm run lint",
+    "precommit":"npm run test",
+    "prepublish": "npm run build",
     "build": "rimraf dist && tsc",
-    "lint": "eslint . --c .ts --c .mts",
-    "test": "jest"
+    "lint": "eslint .",
+    "test": "jest --coverage" 
   },
   "devDependencies": {
+    "@eslint/js": "^9.6.0",
     "@types/aws-lambda": "^8.10.140",
+    "@types/eslint__js": "^8.42.3",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.14.9",
-    "@typescript-eslint/parser": "^7.14.1",
     "eslint": "^9.6.0",
     "eslint-plugin-jest": "^28.6.0",
     "jest": "^29.7.0",
@@ -28,6 +32,7 @@
     "typescript-eslint": "^7.14.1"
   },
   "dependencies": {
+    "ts-node": "^10.9.2",
     "tslib": "^2.6.3"
   }
 }
